+++ Systems/Api/OriginalLanguage.Api/Controllers/Lessons/LessonsController.cs
@@ -10,6 +10,10 @@ using OriginalLanguage.Services.TaskGenerator;
 using OriginalLanguage.Services.TaskAnswerChecker.Models;
 using OriginalLanguage.Services.TaskAnswerChecker;
 using OriginalLanguage.Services.LessonCompletion;
+using OriginalLanguage.Api.Helpers;
+using Microsoft.AspNetCore.Authorization;
+using OriginalLanguage.Consts;
+using OriginalLanguage.Context.Entities;
 
 namespace OriginalLanguage.Api.Controllers.Lessons;
 
@@ -22,14 +26,15 @@ namespace OriginalLanguage.Api.Controllers.Lessons;
 [ApiController]
 [Produces("application/json")]
 [ApiVersion("1.0")]
-public class LessonsController : ControllerBase
+public class LessonsController : AppControllerBase
 {
+    private readonly IMapper mapper;
     private readonly ILessonsService lessonsService;
     private readonly ILessonSamplesService lessonSamplesService;
     private readonly ITaskGenerator taskGenerator;
     private readonly ITaskAnswerChecker taskAnswerChecker;
     private readonly ILessonCompletionService lessonCompletionService;
-    private readonly IMapper mapper;
+    private readonly ResourceOwningHelper resourceOwningHelper;
 
     public LessonsController(
         IMapper mapper,
@@ -37,35 +42,41 @@ public class LessonsController : ControllerBase
         ILessonSamplesService lessonSamplesService,
         ITaskGenerator taskGenerator,
         ITaskAnswerChecker taskAnswerChecker,
-        ILessonCompletionService lessonCompletionService)
+        ILessonCompletionService lessonCompletionService,
+        ResourceOwningHelper resourceOwningHelper,
+        IAuthorizationService authorizationService)
+        : base(authorizationService)
     {
         this.lessonsService = lessonsService;
         this.lessonSamplesService = lessonSamplesService;
         this.taskGenerator = taskGenerator;
         this.taskAnswerChecker = taskAnswerChecker;
         this.lessonCompletionService = lessonCompletionService;
+        this.resourceOwningHelper = resourceOwningHelper;
         this.mapper = mapper;
     }
 
-    [HttpGet("{id}")]
+    
     [ProducesResponseType(typeof(LessonResponse), 200)]
+    [HttpGet("{id}")]
     public async Task<LessonResponse> GetLesson([FromRoute] int id)
     {
         return mapper.Map<LessonResponse>(await lessonsService.GetLesson(id));
     }
 
-    [HttpGet("")]
-    [ProducesResponseType(typeof(IEnumerable<LessonResponse>), 200)]
-    public async Task<IEnumerable<LessonResponse>> GetLessons(
-        [FromQuery] int offset = 0,
-        [FromQuery] int limit = 10)
-    {
-        return mapper.Map<IEnumerable<LessonResponse>>(
-            await lessonsService.GetLessons(offset, limit));
-    }
+    //[ProducesResponseType(typeof(IEnumerable<LessonResponse>), 200)]
+    //[HttpGet("")]
+    //public async Task<IEnumerable<LessonResponse>> GetLessons(
+    //    [FromQuery] int offset = 0,
+    //    [FromQuery] int limit = 10)
+    //{
+    //    return mapper.Map<IEnumerable<LessonResponse>>(
+    //        await lessonsService.GetLessons(offset, limit));
+    //}
 
-    [HttpGet("{lessonId}/lesson-samples")]
     [ProducesResponseType(typeof(IEnumerable<LessonSampleResponse>), 200)]
+    [Authorize(AppScopes.CoursesLearn)]
+    [HttpGet("{lessonId}/lesson-samples")]
     public async Task<IEnumerable<LessonSampleResponse>> GetSamplesOfLesson(
         [FromRoute] int lessonId)
     {
@@ -74,40 +85,64 @@ public class LessonsController : ControllerBase
         return mapper.Map<IEnumerable<LessonSampleResponse>>(lessonSamples);
     }
 
-    [HttpPost("")]
     [ProducesResponseType(typeof(LessonResponse), 200)]
-    public async Task<LessonResponse> AddLesson([FromBody] AddLessonRequest request)
+    [Authorize(AppScopes.ContentWrite)]
+    [HttpPost("")]
+    public async Task<IActionResult> AddLesson(
+        [FromBody] AddLessonRequest request)
     {
-        return mapper.Map<LessonResponse>(
-            await lessonsService.AddLesson(mapper.Map<AddLessonModel>(request)));
+        string ownerId = (await resourceOwningHelper
+            .GetCourseAuthorId(request.CourseId))
+            .ToString();
+        var res = await ForbidIfResourceIsNotOwned(ownerId);
+        if (res != null) return res;
+
+        var addedLessonModel = await lessonsService
+            .AddLesson(mapper.Map<AddLessonModel>(request));
+        return Ok(mapper.Map<LessonResponse>(addedLessonModel));
     }
 
+    [Authorize(AppScopes.ContentWrite)]
     [HttpPut("{id}")]
     public async Task<IActionResult> UpdateLesson(
         [FromRoute] int id,
         [FromBody] UpdateLessonRequest request)
     {
+        var res = await ForbidExistingNotOwnedLesson(id);
+        if (res != null) return res;
+
         await lessonsService.UpdateLesson(id, mapper.Map<UpdateLessonModel>(request));
         return Ok();
     }
 
+    [Authorize(AppScopes.ContentWrite)]
     [HttpPut("numbers")]
     public async Task<IActionResult> UpdateLessonNumber(
         [FromBody] IEnumerable<LessonIdAndNumber> lessonIdsAndNumbers)
     {
+        foreach (var idAndNumber in lessonIdsAndNumbers) {
+            var res = await ForbidExistingNotOwnedLesson(idAndNumber.Id);
+            if (res != null) return res;
+        }
+
         await lessonsService.UpdateLessonNumbers(lessonIdsAndNumbers);
         return Ok();
     }
 
+    [Authorize(AppScopes.ContentWrite)]
     [HttpDelete("{id}")]
     public async Task<IActionResult> DeleteLesson([FromRoute] int id)
     {
+        var res = await ForbidExistingNotOwnedLesson(id);
+        if (res != null) return res;
+
         await lessonsService.DeleteLesson(id);
         return Ok();
     }
 
-    [HttpGet("{id}/generate-tasks")]
     [ProducesResponseType(typeof(IEnumerable<LessonTask>), 200)]
+    [Authorize(AppScopes.CoursesLearn)]
+    [HttpGet("{id}/generate-tasks")]
     public async Task<IEnumerable<LessonTask>> GenerateLessons(
         [FromRoute] int id)
     {
@@ -115,20 +150,30 @@ public class LessonsController : ControllerBase
         return await taskGenerator.GenerateLessonTasks(id, 10);
     }
 
-    [HttpPost("check-task-answer")]
     [ProducesResponseType(typeof(CheckAnswerResult), 200)]
+    [Authorize(AppScopes.CoursesLearn)]
+    [HttpPost("check-task-answer")]
     public async Task<CheckAnswerResult> CheckTaskAnswer(
         [FromBody] TaskAnswer answer)
     {
         return await taskAnswerChecker.Check(answer);
     }
 
-    [HttpPost("{id}/complete")]
     [ProducesResponseType(typeof(LessonCompletionResult), 200)]
+    [Authorize(AppScopes.CoursesLearn)]
+    [HttpPost("{id}/complete")]
     public async Task<LessonCompletionResult> CheckTaskAnswer(
         [FromRoute] int lessonId,
         [FromBody] IEnumerable<TaskAnswer> answers)
     {
         return await lessonCompletionService.TryCompleteLesson(lessonId, answers);
+    }
+
+    private async Task<IActionResult?> ForbidExistingNotOwnedLesson(int lessonId)
+    {
+        string resourceOwnerId = (await resourceOwningHelper
+            .GetLessonAuthorId(lessonId))
+            .ToString();
+        return await ForbidIfResourceIsNotOwned(resourceOwnerId);
     }
 }
